fix: close t8 qa1 direct findings
