+++ crates/atm-daemon/src/test_support.rs
@@ -81,8 +81,25 @@ pub(crate) fn connect_daemon_local_ipc_until_ready(
     ready_rx
         .recv_timeout(std::time::Duration::from_secs(3))
         .expect("daemon local ipc ready signal");
-    LocalSocketStream::connect(
-        atm_core::protocol::daemon_local_ipc_name_from_path(endpoint_path).expect("ipc name"),
+    let ipc_name = atm_core::protocol::daemon_local_ipc_name_from_path(endpoint_path)
+        .expect("ipc name");
+    let deadline = std::time::Instant::now() + std::time::Duration::from_secs(2);
+    let mut attempts = 0usize;
+    let mut last_error = None;
+    while std::time::Instant::now() < deadline {
+        match LocalSocketStream::connect(ipc_name.clone()) {
+            Ok(stream) => return stream,
+            Err(error) => {
+                attempts += 1;
+                last_error = Some(error);
+                std::thread::sleep(std::time::Duration::from_millis(10));
+            }
+        }
+    }
+    panic!(
+        "connect daemon local ipc after ready signal failed after {attempts} attempts: {}",
+        last_error
+            .map(|error| error.to_string())
+            .unwrap_or_else(|| "unknown connect error".to_string())
     )
-    .unwrap_or_else(|error| panic!("connect daemon local ipc after ready signal: {error}"))
 }
