+++ crates/atm-daemon/src/tests_graft.rs
@@ -8,6 +8,7 @@ use atm_core::schema::{AgentMember, TeamConfig};
 use atm_core::test_support::ROLE_TEAM_LEAD;
 use atm_core::types::IsoTimestamp;
 use atm_rusqlite::assemble_boundary;
+use serial_test::serial;
 use tempfile::TempDir;
 
 use crate::runtime_health::{DaemonRequestDispatcher, RuntimeStatusCache};
@@ -72,6 +73,7 @@ fn graft_test_dispatcher() -> (TempDir, DaemonRequestDispatcher) {
 }
 
 #[test]
+#[serial]
 fn dispatcher_routes_graft_register_requests() {
     let (_tempdir, dispatcher) = graft_test_dispatcher();
     let request = graft_registration_request("session-register");
@@ -92,6 +94,7 @@ fn dispatcher_routes_graft_register_requests() {
 }
 
 #[test]
+#[serial]
 fn dispatcher_routes_graft_unregister_requests() {
     let (_tempdir, dispatcher) = graft_test_dispatcher();
     let request = graft_registration_request("session-unregister");
@@ -117,6 +120,7 @@ fn dispatcher_routes_graft_unregister_requests() {
 }
 
 #[test]
+#[serial]
 fn dispatcher_routes_graft_fetch_requests() {
     let (_tempdir, dispatcher) = graft_test_dispatcher();
     let request = graft_registration_request("session-fetch");
@@ -143,6 +147,7 @@ fn dispatcher_routes_graft_fetch_requests() {
 }
 
 #[test]
+#[serial]
 fn dispatcher_routes_graft_drain_requests() {
     let (_tempdir, dispatcher) = graft_test_dispatcher();
     let request = graft_registration_request("session-drain");
